#83 restructuring applications
