+++ .gitignore
@@ -4,8 +4,8 @@
 /cpp/buildArtifact/openSCENARIO*
 /cpp/CMakeLists.txt.user
 /cpp/applications/openScenarioReader/res/result.txt
-/cpp/applications/openScenarioTester/v1_0/src/openScenarioAPI
-/cpp/applications/openScenarioTester/v1_0/src/findHeaders.sh
+/cpp/applications/openScenarioTester/src/openScenarioAPI
+/cpp/applications/openScenarioTester/src/findHeaders.sh
 /cpp/dirs_all.txt
 /cpp/README_template.md
 /cpp/RabbitVCSwithGitHub
@@ -16,6 +16,6 @@
 /doc/allInOne.html
 /cpp/osc.xcworkspace
 /cpp/applications/indexerTester/rc/IndexerTester.aps
-/cpp/applications/openScenarioReader/v1_0/rc/OpenScenarioReader.aps
-/cpp/applications/openScenarioTester/v1_0/rc/OpenScenarioTester.aps
-/cpp/openScenarioLib/v1_0/rc/OpenScenarioLib.aps
+/cpp/applications/openScenarioReader/rc/OpenScenarioReader.aps
+/cpp/applications/openScenarioTester/rc/OpenScenarioTester.aps
+/cpp/openScenarioLib/rc/OpenScenarioLib.aps
