+++ cpp/applications/openScenarioReader/productInfo.txt
@@ -1,0 +1,3 @@
+PRODUCT_NAME=OpenScenarioReader
+PRODUCT_TYPE=exe
+FILE_DESCRIPTION=
