+++ cpp/applications/openScenarioReader/rc/OpenScenarioReader.rc
