--- cpp/applications/openScenarioReader/v1_0/productInfo.txt
@@ -1,3 +1,0 @@
-PRODUCT_NAME=OpenScenarioReader
-PRODUCT_TYPE=exe
-FILE_DESCRIPTION=
