+++ cpp/applications/openScenarioTester/productInfo.txt
@@ -1,0 +1,3 @@
+PRODUCT_NAME=OpenScenarioTester
+PRODUCT_TYPE=exe
+FILE_DESCRIPTION=
