+++ cpp/applications/openScenarioTester/rc/OpenScenarioTester.rc
