+++ cpp/applications/openScenarioTester/src/TestInjectedParameters.h
