+++ cpp/applications/openScenarioTester/src/TestVersionChecker.h
