+++ cpp/applications/openScenarioTester/src/TestWriterApi.h
