+++ cpp/applications/openScenarioTester/src/helper/EgoCheckerRule.h
