+++ cpp/buildArtifact/collectHeaderFiles.sh
@@ -18,7 +18,7 @@ OPEN_SCEANARIO_API=$1
 ################################################################
 # path and file name for findHeaders.sh
 # it is placed in the openScenarioTester project folder
-FIND_HEADERS_SH_PATH=./applications/openScenarioTester/v1_0/src
+FIND_HEADERS_SH_PATH=./applications/openScenarioTester/src
 FIND_HEADERS_SH_FILE=findHeaders.sh
 FIND_HEADERS_SH=${FIND_HEADERS_SH_PATH}/${FIND_HEADERS_SH_FILE}
 
@@ -35,9 +35,9 @@ echo OpenScenarioTester.cpp >> ${FIND_HEADERS_SH}
 chmod a+x ${FIND_HEADERS_SH}
 
 # go to the openScenarioTester folder where its main file OpenScenarioTester.cpp is located
-cd ${SCRIPT_DIR}/../applications/openScenarioTester/v1_0/src/
+cd ${SCRIPT_DIR}/../applications/openScenarioTester/src/
 # prepare the openScenario source for install folder; ok it is a bit clumsy but it works as expected
-echo "rm -rf ${SCRIPT_DIR}/../applications/openScenarioTester/v1_0/src/${OPEN_SCEANARIO_API}"
+echo "rm -rf ${SCRIPT_DIR}/../applications/openScenarioTester/src/${OPEN_SCEANARIO_API}"
 rm -rf ${OPEN_SCEANARIO_API}
 mkdir -p ${OPEN_SCEANARIO_API}/include/a/b/c/d
 # and now let the compiler collect all necessary dependent header files
@@ -49,11 +49,11 @@ for i in `./${FIND_HEADERS_SH_FILE}` ; do
     fi ;
 done
 # clean up
-echo "rm -rf ${SCRIPT_DIR}/../applications/openScenarioTester/v1_0/src/${OPEN_SCEANARIO_API}/include/a"
+echo "rm -rf ${SCRIPT_DIR}/../applications/openScenarioTester/src/${OPEN_SCEANARIO_API}/include/a"
 rm -rf ${OPEN_SCEANARIO_API}/include/a
 
 # move to SCRIPT_DIR but before delete old one
 echo "rm -rf ${SCRIPT_DIR}/${OPEN_SCEANARIO_API}"
-echo "mv ${SCRIPT_DIR}/../applications/openScenarioTester/v1_0/src/${OPEN_SCEANARIO_API} ${SCRIPT_DIR}"
+echo "mv ${SCRIPT_DIR}/../applications/openScenarioTester/src/${OPEN_SCEANARIO_API} ${SCRIPT_DIR}"
 rm -rf ${SCRIPT_DIR}/${OPEN_SCEANARIO_API}
-mv ${SCRIPT_DIR}/../applications/openScenarioTester/v1_0/src/${OPEN_SCEANARIO_API} ${SCRIPT_DIR}
+mv ${SCRIPT_DIR}/../applications/openScenarioTester/src/${OPEN_SCEANARIO_API} ${SCRIPT_DIR}
